CrissCrossDHT (criss_cross_dht v0.0.1)

CrissCrossDHT is an Elixir package that provides a Kademlia Distributed Hash Table (DHT) implementation according to BitTorrent Enhancement Proposals (BEP) 05. This specific implementation is called "mainline" variant.

Types

cluster()

Specs

cluster() :: binary()

A binary which contains the cluster

infohash()

Specs

infohash() :: binary()

A binary which contains the infohash of a torrent. An infohash is a SHA1 encoded hex sum which identifies a torrent.

node_id()

Specs

node_id() :: <<_::34>>

TODO

node_id_enc()

Specs

node_id_enc() :: String.t()

TODO

tcp_port()

Specs

tcp_port() :: 0..65565

A non negative integer (0--65565) which represents a TCP port number.

ttl()

Specs

ttl() :: integer()

Functions

cluster_announce(cluster, infohash, ttl)

find_name(cluster, infohash, generation, callback)

find_name_sync(cluster, infohash, generation, timeout \\ 10000)

find_value(cluster, infohash, callback)

find_value_sync(cluster, infohash, timeout \\ 10000)

has_announced(cluster, infohash)

node_id()

Specs

node_id() :: node_id()

This function returns the generated node_id as a bitstring.

node_id_enc()

Specs

node_id_enc() :: node_id_enc()

This function returns the generated node_id encoded as a String (40 characters).

ref()

search(cluster, infohash, callback)

Specs

search(cluster(), infohash(), (... -> any())) :: atom()

This function needs an infohash as binary and a callback function as parameter. This function uses its own routing table as a starting point to start a get_peers search for the given infohash.

Example

iex> "3F19B149F53A50E14FC0B79926A391896EABAB6F"
      |> Base.decode16!
      |> CrissCrossDHT.search(fn(node) ->
       {ip, port} = node
       IO.puts "ip: #{inspect ip} port: #{port}"
     end)

search_announce(cluster, infohash, ttl, callback)

Specs

search_announce(cluster(), infohash(), ttl(), (... -> any())) :: atom()

This function needs an infohash as binary and callback function as parameter. This function does the same thing as the search/2 function, except it sends an announce message to the found peers. This function does not need a TCP port which means the announce message sets :implied_port to true.

Example

iex> "3F19B149F53A50E14FC0B79926A391896EABAB6F"
     |> Base.decode16!
     |> CrissCrossDHT.search_announce(fn(node) ->
       {ip, port} = node
       IO.puts "ip: #{inspect ip} port: #{port}"
     end)

search_announce(cluster, infohash, callback, ttl, port)

Specs

search_announce(cluster(), infohash(), (... -> any()), ttl(), tcp_port()) ::
  atom()

This function needs an infohash as binary, a callback function as parameter, and a TCP port as integer. This function does the same thing as the search/2 function, except it sends an announce message to the found peers.

Example

iex> "3F19B149F53A50E14FC0B79926A391896EABAB6F" ## Ubuntu 15.04
     |> Base.decode16!
     |> CrissCrossDHT.search_announce(fn(node) ->
       {ip, port} = node
       IO.puts "ip: #{inspect ip} port: #{port}"
     end, 6881)

store(cluster, value, ttl)

store(cluster, value, ttl, tid)

store_name(cluster, rsa_priv_key, value, save_local, save_remote, ttl)

store_name(cluster, rsa_priv_key, value, save_local, save_remote, ttl, tid)